新規登録専用のコントローラの作成及び分離
