+++ src/main/java/oit/is/hondaken/scheduler/controller/ScheduleController.java
@@ -45,13 +45,6 @@ import oit.is.hondaken.scheduler.model.Week;
 import oit.is.hondaken.scheduler.model.Week2;
 import oit.is.hondaken.scheduler.service.TimeTableService;
 
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-// import org.springframework.mail.MailSender;
-// import org.springframework.mail.SimpleMailMessage;
-
-import java.util.Random;
-
 @Controller
 public class ScheduleController {
   private static final Logger logger = LoggerFactory.getLogger(ScheduleController.class);
@@ -315,75 +308,6 @@ public class ScheduleController {
     redirectAttributes.addFlashAttribute("message", scheduleMapper.selectClassNameById(classId) + message);
 
     return "redirect:/timetable";
-  }
-
-  @GetMapping("/register")
-  public String goreg() {
-    return "register.html";
-  }
-
-  @GetMapping("/admin")
-  public String sending() {
-    return "codechk.html";
-  }
-
-  @PostMapping("/admin")
-  public String codechk(@RequestParam String gakuseki, @RequestParam String mail, @RequestParam String pass,
-      @RequestParam String myname,
-      ModelMap model) {
-    Random rnd = new Random();
-    ArrayList<String> Numbers = userSettingMapper.selectNumber();
-    ArrayList<String> Mails = userSettingMapper.selectMail();
-    String admin_code = rnd.nextInt(1000000) + "";
-
-    int flag = 0;
-    for (String num : Numbers) {
-      if (gakuseki.equals(num)) {
-        flag = 1;
-      }
-    }
-    for (String addr : Mails) {
-      if (mail.equals(addr)) {
-        flag = 2;
-      }
-    }
-    if (flag == 0) {// ここからメール送信
-      // SimpleMailMessage message = new SimpleMailMessage();
-      // message.setTo(mail);
-      // message.setFrom("");
-      // // こことaporication propatoiesのメアドの入力を忘れないようにすること！
-      // message.setSubject("認証コードをお送りします。");
-      // message.setText("認証コード:" + admin_code);
-      // // メール送信を実施する。
-      // mailSender.send(message);
-      model.addAttribute("mail", mail);
-      model.addAttribute("gakuseki", gakuseki);
-      model.addAttribute("pass", pass);
-      model.addAttribute("myname", myname);
-      model.addAttribute("admin_code", admin_code);
-    }
-    model.addAttribute("flag", flag);
-
-    return "codechk.html";
-  }
-
-  @PostMapping("/regfin")
-  public String regfin(@RequestParam String gakuseki, @RequestParam String mail, @RequestParam String pass,
-      @RequestParam String myname,
-      ModelMap model) {
-
-    UserSetting user = new UserSetting();
-    user.setMyNumber(gakuseki);
-    user.setUserName(myname);
-    user.setMail(mail);
-    BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-    String hashedPassword = encoder.encode(pass);
-    user.setMyPass(hashedPassword);
-    user.setUserRole("STUDENT");
-    user.setActive(false);
-    userSettingMapper.insertuserSetting(user);
-    model.addAttribute("user", user);
-    return "regfin.html";
   }
 
   @GetMapping("/todolist")
