+++ src/main/resources/application.properties
@@ -14,3 +14,11 @@ spring.h2.console.enabled:true
 logging.level.root=INFO
 logging.level.org.springframework=DEBUG
 logging.level.oit.is.hondaken=DEBUG
+
+spring.mail.host=localhost
+spring.mail.port=25
+spring.mail.username=
+spring.mail.password=
+spring.mail.protocol=smtp
+spring.mail.properties.mail.smtp.auth=false
+spring.mail.properties.mail.smtp.starttls.enable=false
